pyproject.toml
[tool.poetry]
name = "languagecreationtools"
version = "1.0"
description = "Some tools which can help one to create one's own language, featuring different aspects of language."
authors = ["cwellm <christoph.wellm@creaflect.de>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "==6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
